pyproject.toml
[project]
name = "travel-agent"
version = "0.1.0"
description = "LLM agent with retrieval"
readme = "README.md"
authors = [
    { name = "Тот самый белый", email = "kirazu@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "accelerate>=1.6.0",
    "fastembed-gpu>=0.6.1",
    "huggingface-hub[hf-xet]>=0.30.2",
    "loguru>=0.7.3",
    "pandas>=2.2.3",
    "pydantic>=2.11.3",
    "pymystem3>=0.2.0",
    "python-dotenv>=1.1.0",
    "qdrant-client>=1.13.3",
    "sentence-transformers>=4.0.2",
    "smolagents[gradio,litellm,telemetry]>=1.13.0",
    "torch>=2.6.0",
    "tqdm>=4.67.1",
    "transformers>=4.51.2",
    "vertexai>=1.71.1",
]

[dependency-groups]
dev = [
    "jupyter>=1.1.1",
    "matplotlib>=3.10.1",
    "nltk>=3.9.1",
    "numpy>=2.2.4",
    "plotly>=6.0.1",
    "pre-commit>=4.2.0",
    "ruff>=0.11.5",
    "seaborn>=0.13.2",
    "transliterate>=1.10.2",
    "umap-learn>=0.5.7",
    "wordcloud>=1.9.4",
]
docs = [
    "mkdocs-material>=9.6.11",
]
langchain = [
    "langchain-ollama>=0.3.2",
    "langchain-qdrant>=0.2.0",
    "langchain>=0.3.24",
]

[project.scripts]
travel-agent = "travel_agent:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.isort]
profile = "black"
skip_gitignore = true
line_length = 120
float_to_top = true
known_third_party = ["wandb"]

[tool.ruff]
target-version = "py313"
line-length = 120

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F", "I"]
